// src/types/generated/ConnectionStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ConnectionState } from "./ConnectionState";

export type ConnectionStatus = { serverId: string, state: ConnectionState, latencyMs: number | null, fingerprint: string | null, serverBanner: string | null, };